refactor: #1187: Renamed classes and moved settings to more suitable namespace.
