+++ Atlas.Debug.Client.Models/Settings/MatchingAlgorithmHttpFunctionSettings.cs
@@ -1,6 +1,4 @@
-﻿using Atlas.Debug.Client.Models.HttpFunctions;
-
-namespace Atlas.Debug.Client.Models.MatchingAlgorithm
+﻿namespace Atlas.Debug.Client.Models.Settings
 {
     /// <summary>
     /// Settings for calling the matching algorithm debug functions.
