+++ Atlas.Debug.Client.Models/Settings/TopLevelHttpFunctionSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace Atlas.Debug.Client.Models.Settings
+{
+    /// <summary>
+    /// Settings for calling the support message debug functions.
+    /// </summary>
+    public class TopLevelHttpFunctionSettings : HttpFunctionSettings
+    {
+    }
+}
