--- Atlas.Debug.Client.Models/SupportMessages/SupportMessageHttpFunctionSettings.cs
@@ -1,11 +1,0 @@
-﻿using Atlas.Debug.Client.Models.HttpFunctions;
-
-namespace Atlas.Debug.Client.Models.SupportMessages
-{
-    /// <summary>
-    /// Settings for calling the support message debug functions.
-    /// </summary>
-    public class SupportMessageHttpFunctionSettings : HttpFunctionSettings
-    {
-    }
-}
