+++ Atlas.Debug.Client/Clients/MatchingAlgorithmFunctionsClient.cs
@@ -8,7 +8,7 @@ namespace Atlas.Debug.Client.Clients
     /// <summary>
     /// Client for calling matching algorithm-related debug functions.
     /// </summary>
-    public interface IMatchingAlgorithmClient
+    public interface IMatchingAlgorithmFunctionsClient
     {
         /// <summary>
         /// Check for presence or absence of donors in the active copy of the matching algorithm database.
@@ -16,11 +16,11 @@ namespace Atlas.Debug.Client.Clients
         Task<DebugDonorsResult> CheckDonors(IEnumerable<string> externalDonorCodes);
     }
 
-    /// <inheritdoc cref="IMatchingAlgorithmClient" />
-    public class MatchingAlgorithmClient : HttpFunctionClient, IMatchingAlgorithmClient
+    /// <inheritdoc cref="IMatchingAlgorithmFunctionsClient" />
+    public class MatchingAlgorithmFunctionsClient : HttpFunctionClient, IMatchingAlgorithmFunctionsClient
     {
         /// <inheritdoc />
-        public MatchingAlgorithmClient(HttpClient client) : base(client)
+        public MatchingAlgorithmFunctionsClient(HttpClient client) : base(client)
         {
         }
